image_analysis/ConvexiPy/requirements.txt
opencv-python
pathlib
Pillow
scipy
numpy
matplotlib
scikit-image